requirements.txt
anyio==3.7.1
fastapi==0.75.2
idna==3.11
pydantic==1.10.26
sniffio==1.3.1
SQLAlchemy==1.4.36
starlette==0.17.1
typing_extensions==4.15.0
alembic==1.7.7
psycopg2-binary
starlette-context==0.3.4
typeguard
loguru
PyJWT==2.6.0
passlib
